requirements.txt
# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# COCO Evaluation Tools
pycocotools>=2.0.6

# Configuration
PyYAML>=6.0

# Progress Bars
tqdm>=4.65.0

# Visualization
matplotlib>=3.7.0

# Machine Learning Utilities
scikit-learn>=1.3.0

# TensorBoard for Logging
tensorboard>=2.13.0


